requirements.txt
psutil
pywin32
opencv-python
Pillow
soundcard
numpy
moviepy
